+++ okpt/test/steps/nmslib.py
@@ -1,5 +1,5 @@
 import numpy as np
-from okpt.io.config.parsers.nmslib import NmslibConfig
+from okpt.io.config.parsers import nmslib as nmslib_parser
 from okpt.test.profile import label, measure
 
 import nmslib
@@ -7,7 +7,7 @@ import nmslib
 
 @label('init_index')
 @measure
-def init_index(service_config: NmslibConfig):
+def init_index(service_config: nmslib_parser.NmslibConfig):
     index = nmslib.init(method=service_config.method.name,
                         space=service_config.method.space_type)
     return {'index': index}
@@ -22,7 +22,8 @@ def bulk_index(index: nmslib.dist.FloatIndex, dataset: np.ndarray):
 
 @label('create_index')
 @measure
-def create_index(index: nmslib.dist.FloatIndex, service_config: NmslibConfig):
+def create_index(index: nmslib.dist.FloatIndex,
+                 service_config: nmslib_parser.NmslibConfig):
     index.createIndex({
         'efConstruction': service_config.method.parameters.ef_construction,
         'M': service_config.method.parameters.M,
