+++ okpt/test/steps/opensearch.py
@@ -1,9 +1,9 @@
 from typing import Any, Dict, Iterable
 
+import elasticsearch
 import h5py
 import numpy as np
-from elasticsearch import Elasticsearch
-from okpt.test.profile import label, measure
+from okpt.test import profile
 
 
 def bulk_partition_vectors(dataset: h5py.Dataset, split_size: int):
@@ -23,24 +23,24 @@ def bulk_transform_vectors(dataset: h5py.Dataset, action: Dict[str, Any],
     return sections
 
 
-@label('create_index')
-@measure
-def create_index(es: Elasticsearch, index_name: str, index_spec: Dict[str,
-                                                                      Any]):
+@profile.label('create_index')
+@profile.measure
+def create_index(es: elasticsearch.Elasticsearch, index_name: str,
+                 index_spec: Dict[str, Any]):
     return es.indices.create(index=index_name, body=index_spec)
 
 
-@label('disable_refresh')
-@measure
-def disable_refresh(es: Elasticsearch):
+@profile.label('disable_refresh')
+@profile.measure
+def disable_refresh(es: elasticsearch.Elasticsearch):
     return es.indices.put_settings(body={'index': {'refresh_interval': -1}})
 
 
-def bulk_index(es: Elasticsearch, index_name: str,
+def bulk_index(es: elasticsearch.Elasticsearch, index_name: str,
                sections: Iterable[np.ndarray]):
-    @label('bulk_add')
+    @profile.label('bulk_add')
     def bulk(index, body):
-        return es.bulk(index=index, body=body, timeout='60s')
+        return es.bulk(index=index, body=body)
 
     results = []
     for section in sections:
@@ -49,16 +49,17 @@ def bulk_index(es: Elasticsearch, index_name: str,
     return results
 
 
-@label('refresh_index')
-@measure
-def refresh_index(es: Elasticsearch, index_name: str):
+@profile.label('refresh_index')
+@profile.measure
+def refresh_index(es: elasticsearch.Elasticsearch, index_name: str):
     return es.indices.refresh(index=index_name)
 
 
-def delete_index(es: Elasticsearch, index_name: str):
+def delete_index(es: elasticsearch.Elasticsearch, index_name: str):
     es.indices.delete(index=index_name)
 
 
-@label('query_index')
-def query_index(es: Elasticsearch, index_name: str, body: Dict[str, Any]):
+@profile.label('query_index')
+def query_index(es: elasticsearch.Elasticsearch, index_name: str,
+                body: Dict[str, Any]):
     return es.search(index=index_name, body=body)
